+++ react/DrawerCloseButton.tsx
@@ -1,0 +1,29 @@
+import React from 'react'
+import { useCssHandles } from 'vtex.css-handles'
+import { IconClose } from 'vtex.store-icons'
+
+import { useDrawer } from './DrawerContext'
+
+const CSS_HANDLES = ['closeIconButton'] as const
+
+interface Props {
+  size?: number
+  type?: 'filled' | 'line'
+}
+
+const DrawerCloseButton: React.FC<Props> = ({ size = 30, type = 'line' }) => {
+  const { close } = useDrawer()
+
+  const handles = useCssHandles(CSS_HANDLES)
+
+  return (
+    <button
+      className={`${handles.closeIconButton} pa4 pointer bg-transparent transparent bn pointer`}
+      onClick={close}
+    >
+      <IconClose size={size} type={type} />
+    </button>
+  )
+}
+
+export default DrawerCloseButton