+++ react/typings/vtex.render-runtime.d.ts
@@ -1,0 +1,7 @@
+/* eslint-disable import/order */
+
+import * as Runtime from 'vtex.render-runtime'
+
+declare module 'vtex.render-runtime' {
+  export const useChildBlock: (opts: { id: string }) => {} | null
+}