+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.10.0] - 2020-03-30
+### Added
+- New CSS Handle `drawerContent`
+
+## [0.9.0] - 2020-03-13
+### Added
+- New block `drawer-header` and `drawer-close-button`.
+
+### Fixed
+- Drawer trigger block never being rendered.
+
 ## [0.8.1] - 2020-03-11
 ### Fixed
 - Correctly identify if animation is stopped or not.